refactor: Main 컴포넌트 추가, Layout 분리
